Remove the multiply inherited mix-in classes from the states.
This changes the implementation of the TReconState derived objects from
using multiply inherited mix-in states which define the interface to a
"old-style" C implementation using CPP definitions.  The old-style
implementation plays better with the ROOT i/o implementation, even if it is
not as elegant (or as easy to maintain).